Server's message handler can return arbitrary container of data now
It had to return an std::vector before, but it's pretty prohibitive
when interfacing with another programming language.
The programmer would have to either copy the data to a vector or write
some wrapper around it.
Now the only restriction is std::size() and std::data() working on the
container provided by the programmer.